Moved some more functions to util
