+++ src/camadautil.h
@@ -2,7 +2,9 @@
 #define CAMADAUTIL_H_
 
 #include <bitset>
+#include <cassert>
 #include <cstdint>
+#include <cstring>
 #include <string>
 
 namespace camada {
@@ -15,6 +17,26 @@ static inline std::string toTwosComplementBin(int64_t Value, unsigned Width) {
   else if (Width > 64)
     Bits.insert(Bits.begin(), Width - 64, Value < 0 ? '1' : '0');
   return Bits;
+}
+
+template <typename FPType, typename IntType>
+static inline FPType IntAsFP(const IntType Int) {
+  assert(sizeof(FPType) == sizeof(IntType) &&
+         "Cannot convert int to floating-point");
+
+  FPType Result;
+  std::memcpy(&Result, &Int, sizeof(IntType));
+  return Result;
+}
+
+template <typename FPType, typename IntType>
+static inline IntType FPAsInt(const FPType FP) {
+  assert(sizeof(FPType) == sizeof(IntType) &&
+         "Cannot convert int to floating-point");
+
+  IntType Result;
+  std::memcpy(&Result, &FP, sizeof(FPType));
+  return Result;
 }
 
 } // namespace camada
